requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
groq==0.9.0
httpx==0.27.0
google-auth==2.29.0
google-api-python-client==2.128.0
audioop-lts==0.2.1   # audioop for Python 3.13+; use built-in audioop for Python 3.12 and below
